requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0=
colorama==0.4.6
distlib==0.3.7=
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-tinymce==3.6.1
djangorestframework==3.14.0
environ==1.0
filelock==3.12.2
idna==3.4
iso8601==0.1.12
pbr==5.11.1
Pillow==10.0.1
pipenv==2023.7.23
platformdirs==3.10.0
pypng==0.20220715.0
pytermii==0.0.3
pytz==2019.3
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.16
virtualenv==20.24.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
